# nodejs-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is the API for the Estuary application."
  version: "0.0.0"
  title: "Estuary API"
  termsOfService: "http://estuary.tech"
  contact:
    name: "API Support"
    url: "https://docs.estuary.tech/feedback"
  license:
    name: "Apache 2.0 Apache-2.0 OR MIT"
    url: "https://github.com/application-research/estuary/blob/master/LICENSE.md"
host: "api.estuary.tech"
basePath: "/"
security:
- bearerAuth: []
paths:
...
  /content/deals:
    get:
      tags:
      - "content"
      summary: "Content with deals"
      description: "This endpoint lists all content with deals"
      operationId: "contentDealsGET"
      produces:
      - "application/json"
      parameters:
      - name: "limit"
        in: "query"
        description: "Limit"
        required: false
        type: "integer"
      - name: "offset"
        in: "query"
        description: "Offset"
        required: false
        type: "integer"
      responses: {}
      x-swagger-router-controller: "Content"
...
  /deal/estimate:
    post:
      tags:
      - "deals"
      summary: "Estimate the cost of a deal"
      description: "This endpoint estimates the cost of a deal"
      operationId: "dealEstimatePOST"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "The size of the deal in bytes, the replication factor, and the\
          \ duration of the deal in blocks"
        required: true
        schema:
          $ref: "#/definitions/main.estimateDealBody"
      responses: {}
      x-swagger-router-controller: "Deals"
  /deal/info/{dealid}:
    get:
      tags:
      - "deals"
      summary: "Get Deal Info"
      description: "This endpoint returns the deal info for a deal"
      operationId: "dealInfoDealidGET"
      produces:
      - "application/json"
      parameters:
      - name: "dealid"
        in: "path"
        description: "Deal ID"
        required: true
        type: "integer"
      responses: {}
      x-swagger-router-controller: "Deals"
  /deal/proposal/{propcid}:
    get:
      tags:
      - "deals"
      summary: "Get Proposal"
      description: "This endpoint returns the proposal for a deal"
      operationId: "dealProposalPropcidGET"
      produces:
      - "application/json"
      parameters:
      - name: "propcid"
        in: "path"
        description: "Proposal CID"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Deals"
  /deal/query/{miner}:
    get:
      tags:
      - "deals"
      summary: "Query Ask"
      description: "This endpoint returns the ask for a given CID"
      operationId: "dealQueryMinerGET"
      produces:
      - "application/json"
      parameters:
      - name: "miner"
        in: "path"
        description: "CID"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Deals"
  /deal/status-by-proposal/{propcid}:
    get:
      tags:
      - "deals"
      summary: "Get Deal Status by PropCid"
      description: "Get Deal Status by PropCid"
      operationId: "dealStatus_by_proposalPropcidGET"
      produces:
      - "application/json"
      parameters:
      - name: "propcid"
        in: "path"
        description: "PropCid"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Deals"
  /deal/status/{miner}/{propcid}:
    get:
      tags:
      - "deals"
      summary: "Deal Status"
      description: "This endpoint returns the status of a deal"
      operationId: "dealStatusMinerPropcidGET"
      produces:
      - "application/json"
      parameters:
      - name: "miner"
        in: "path"
        description: "Miner"
        required: true
        type: "string"
      - name: "propcid"
        in: "path"
        description: "Proposal CID"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Deals"
...
  /pinning/pins:
    get:
      tags:
      - "pinning"
      summary: "List all pin status objects"
      description: "This endpoint lists all pin status objects"
      operationId: "pinningPinsGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        "400":
          description: "Bad Request"
          schema:
            $ref: "#/definitions/util.HttpError"
        "404":
          description: "Not Found"
          schema:
            $ref: "#/definitions/util.HttpError"
        "500":
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/util.HttpError"
      x-swagger-router-controller: "Pinning"
    post:
      tags:
      - "pinning"
      summary: "Add and pin object"
      description: "This endpoint adds a pin to the IPFS daemon."
      operationId: "pinningPinsPOST"
      produces:
      - "application/json"
      parameters:
      - name: "cid"
        in: "path"
        description: "cid"
        required: true
        type: "string"
      - name: "name"
        in: "path"
        description: "name"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Pinning"
  /pinning/pins/{pinid}:
    get:
      tags:
      - "pinning"
      summary: "Get a pin status object"
      description: "This endpoint returns a pin status object."
      operationId: "pinningPinsPinidGET"
      produces:
      - "application/json"
      parameters:
      - name: "pinid"
        in: "path"
        description: "cid"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Pinning"
    post:
      tags:
      - "pinning"
      summary: "Replace a pinned object"
      description: "This endpoint replaces a pinned object."
      operationId: "pinningPinsPinidPOST"
      produces:
      - "application/json"
      parameters:
      - name: "pinid"
        in: "path"
        description: "Pin ID"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Pinning"
    delete:
      tags:
      - "pinning"
      summary: "Delete a pinned object"
      description: "This endpoint deletes a pinned object."
      operationId: "pinningPinsPinidDELETE"
      produces:
      - "application/json"
      parameters:
      - name: "pinid"
        in: "path"
        description: "Pin ID"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Pinning"
...
  /public/miners/deals/{miner}:
    get:
      tags:
      - "public"
      - "miner"
      summary: "Get all miners deals"
      description: "This endpoint returns all miners deals"
      operationId: "publicMinersDealsMinerGET"
      produces:
      - "application/json"
      parameters:
      - name: "miner"
        in: "path"
        description: "Filter by miner"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Public"
  /public/miners/failures/{miner}:
    get:
      tags:
      - "public"
      - "net"
      summary: "Get all miners"
      description: "This endpoint returns all miners"
      operationId: "publicMinersFailuresMinerGET"
      produces:
      - "application/json"
      parameters:
      - name: "miner"
        in: "path"
        description: "Filter by miner"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Public"
  /public/miners/stats/{miner}:
    get:
      tags:
      - "public"
      - "miner"
      summary: "Get miner stats"
      description: "This endpoint returns miner stats"
      operationId: "publicMinersStatsMinerGET"
      produces:
      - "application/json"
      parameters:
      - name: "miner"
        in: "path"
        description: "Filter by miner"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Public"
  /public/miners/storage/query/{miner}:
    get:
      tags:
      - "deals"
      summary: "Query Ask"
      description: "This endpoint returns the ask for a given CID"
      operationId: "publicMinersStorageQueryMinerGET"
      produces:
      - "application/json"
      parameters:
      - name: "miner"
        in: "path"
        description: "CID"
        required: true
        type: "string"
      responses: {}
      x-swagger-router-controller: "Deals"
...
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  main.estimateDealBody:
    type: "object"
    properties:
      durationBlks:
        type: "integer"
      replication:
        type: "integer"
      size:
        type: "integer"
      verified:
        type: "boolean"
...
  util.HttpError:
    type: "object"
    properties:
      code:
        type: "integer"
      details:
        type: "string"
      reason:
        type: "string"